fix issues with non existing daemon.json
A missing daemon.json is not an error, so should
not show as such in the daemon logs.

Also, the reload logic was flawed, because a
daemon.json could change settings on the daemon,
but once the daemon.json is removed, settings
would never be reloaded, so configuration changes
loaded from daemon.json would not be reverted.

This changes the reload logic, by loading the
configuration file in a separate function,
and reload configuration always - using an
"empty" configuration if the daemon.json was
not found.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>